fix: 프로필 설정 수정 및 위치 이동 profile-setup -> login/profile-setup
